requirements.txt
colorama
requests
pycryptodome
rich
